Add dynamic size

+++ MyDictionary.cpp
@@ -1,7 +1,7 @@
 #include "MyDictionary.h"
 
-const unsigned short int HashFunction(const std::string& key) {
-	unsigned short int hash = 5381;
+const unsigned int HashFunction(const std::string& key) {
+	unsigned int hash = 5381;
 	for (char c : key) {
 		hash = ((hash << 5) + hash) + c;
 	}
